+++ index.js
@@ -1,4 +1,4 @@
-const { rdtsc, setThreadPriority, getThreadPriority, isWin } = require('./build/Release/binding');
+const { rdtsc, setThreadPriority, getThreadPriority, setProcessPriority, getProcessPriority, isWin } = require('./build/Release/binding');
 
 const THREAD_PRIORITY_IDLE = -15;
 const THREAD_PRIORITY_LOWEST = -2;
@@ -6,7 +6,14 @@ const THREAD_PRIORITY_BELOW_NORMAL = -1;
 const THREAD_PRIORITY_NORMAL = 0;
 const THREAD_PRIORITY_ABOVE_NORMAL = 1;
 const THREAD_PRIORITY_HIGHEST = 2;
-const THREAD_PRIORITY_TIME_CRITICAL = 15;
+const THREAD_PRIORITY_REALTIME = 15; // THREAD_PRIORITY_TIME_CRITICAL
+
+const PROCESS_PRIORITY_IDLE = 0x00000040; // IDLE_PRIORITY_CLASS
+const PROCESS_PRIORITY_BELOW_NORMAL = 0x00004000; // BELOW_NORMAL_PRIORITY_CLASS
+const PROCESS_PRIORITY_NORMAL = 0x00000020; // NORMAL_PRIORITY_CLASS
+const PROCESS_PRIORITY_ABOVE_NORMAL = 0x00008000; // ABOVE_NORMAL_PRIORITY_CLASS
+const PROCESS_PRIORITY_HIGHEST = 0x00000080; // HIGH_PRIORITY_CLASS
+const PROCESS_PRIORITY_REALTIME = 0x00000100; // REALTIME_PRIORITY_CLASS
 
 const calcPerformance = function (func0, func1, testTimeMilliseconds) {
     let lastResult;
@@ -14,9 +21,14 @@ const calcPerformance = function (func0, func1, testTimeMilliseconds) {
     let minCycles1 = null;
     let startTime = process.hrtime.bigint();
     let testTime = testTimeMilliseconds * 1000000; //to nano time
-	let previousPriority = setThreadPriority(THREAD_PRIORITY_TIME_CRITICAL);
+	
+	let previousThreadPriority = getThreadPriority();
+	let previousProcessPriority = getProcessPriority();
 	
 	try {
+		setThreadPriority(THREAD_PRIORITY_TIME_CRITICAL);
+		setProcessPriority(PROCESS_PRIORITY_REALTIME);
+
 		let i = 0;
 		do {
 			let cycles0, cycles1;
@@ -43,7 +55,8 @@ const calcPerformance = function (func0, func1, testTimeMilliseconds) {
 			? Number(minCycles1 - minCycles0)
 			: (minCycles1 != null ? minCycles1 : minCycles0);
 	} finally {
-		setThreadPriority(previousPriority);
+		setThreadPriority(previousThreadPriority);
+		setProcessPriority(previousProcessPriority);
 	}
 };
 
@@ -51,13 +64,25 @@ module.exports = {
 	rdtsc,
 	calcPerformance,
 	isWin: isWin(),
+	
 	setThreadPriority, //since only for Windows
 	getThreadPriority, //since only for Windows
+	
 	THREAD_PRIORITY_IDLE,
 	THREAD_PRIORITY_LOWEST,
 	THREAD_PRIORITY_BELOW_NORMAL,
 	THREAD_PRIORITY_NORMAL,
 	THREAD_PRIORITY_ABOVE_NORMAL,
 	THREAD_PRIORITY_HIGHEST,
-	THREAD_PRIORITY_TIME_CRITICAL
+	THREAD_PRIORITY_REALTIME,
+	
+	setProcessPriority, //since only for Windows
+	getProcessPriority, //since only for Windows
+	
+	PROCESS_PRIORITY_IDLE,
+	PROCESS_PRIORITY_BELOW_NORMAL,
+	PROCESS_PRIORITY_NORMAL,
+	PROCESS_PRIORITY_ABOVE_NORMAL,
+	PROCESS_PRIORITY_HIGHEST,
+	PROCESS_PRIORITY_REALTIME
 };