+++ HISTORY.md
@@ -1,3 +1,9 @@
+1.0.2 / 2018-12-02
+===================
+
+  * Added functions: setProcessPriority, getProcessPriority
+  * Improved calcPerformance function
+  
 1.0.1 / 2018-12-02
 ===================